Add pages to MLR UI
