+++ package.json
@@ -12,6 +12,7 @@
     "axios": "^0.19.2",
     "core-js": "^3.6.4",
     "vue": "^2.6.10",
+    "vue-router": "^3.2.0",
     "vuetify": "^2.2.11"
   },
   "devDependencies": {
