+++ src/main.js
@@ -1,6 +1,7 @@
 import Vue from "vue";
-import App from "./App.vue";
-import vuetify from './plugins/vuetify';
+import App from "@/App.vue";
+import vuetify from '@/plugins/vuetify';
+import router from "@/plugins/router";
 import axios from 'axios';
 
 axios.defaults.baseURL = process.env.VUE_APP_API_GATEWAY;
@@ -9,5 +10,6 @@ Vue.config.productionTip = false;
 
 new Vue({
   vuetify,
+  router,
   render: h => h(App)
 }).$mount("#app");
